requirements.txt
Django==1.9
django-crispy-forms==1.6.0
django-filter==0.13.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.8.2
Markdown==2.6.6
markdown2==2.3.5
Pillow==5.0.0
